atc: add flight controller basic implementation
