+++ .gitignore
@@ -11,8 +11,6 @@
 test_output
 build_output
 
-cache
-
 debug.txt
 
 temp
