+++ cmd/atc/resources.go
@@ -14,14 +14,14 @@ import (
 
 	"github.com/yokecd/yoke/internal"
 	"github.com/yokecd/yoke/internal/k8s"
-	"github.com/yokecd/yoke/pkg/apis/airway/v1alpha1"
+	"github.com/yokecd/yoke/pkg/apis/v1alpha1"
 	"github.com/yokecd/yoke/pkg/openapi"
 )
 
 func ApplyResources(ctx context.Context, client *k8s.Client, cfg *Config) error {
 	var (
-		group = "yoke.cd"
-		names = apiextensionsv1.CustomResourceDefinitionNames{
+		group       = "yoke.cd"
+		airwayNames = apiextensionsv1.CustomResourceDefinitionNames{
 			Plural:   "airways",
 			Singular: "airway",
 			Kind:     "Airway",
@@ -38,11 +38,11 @@ func ApplyResources(ctx context.Context, client *k8s.Client, cfg *Config) error 
 			APIVersion: apiextensionsv1.SchemeGroupVersion.Identifier(),
 		},
 		ObjectMeta: metav1.ObjectMeta{
-			Name: names.Plural + "." + group,
+			Name: airwayNames.Plural + "." + group,
 		},
 		Spec: apiextensionsv1.CustomResourceDefinitionSpec{
 			Group: group,
-			Names: names,
+			Names: airwayNames,
 			Scope: apiextensionsv1.ClusterScoped,
 			Versions: []apiextensionsv1.CustomResourceDefinitionVersion{
 				{
@@ -88,7 +88,62 @@ func ApplyResources(ctx context.Context, client *k8s.Client, cfg *Config) error 
 		return fmt.Errorf("failed to convert airway crd to its unstructured representation: %w", err)
 	}
 
-	if err := client.ApplyResource(ctx, airwayResource, forceful); err != nil {
+	flightDef := &apiextensionsv1.CustomResourceDefinition{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "CustomResourceDefinition",
+			APIVersion: apiextensionsv1.SchemeGroupVersion.Identifier(),
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name: "flights.yoke.cd",
+		},
+		Spec: apiextensionsv1.CustomResourceDefinitionSpec{
+			Group: group,
+			Names: apiextensionsv1.CustomResourceDefinitionNames{
+				Plural:   "flights",
+				Singular: "flight",
+				Kind:     "Flight",
+			},
+			Scope: apiextensionsv1.NamespaceScoped,
+			Versions: []apiextensionsv1.CustomResourceDefinitionVersion{
+				{
+					Name:         "v1alpha1",
+					Served:       true,
+					Storage:      true,
+					Subresources: &apiextensionsv1.CustomResourceSubresources{Status: &apiextensionsv1.CustomResourceSubresourceStatus{}},
+					Schema:       &apiextensionsv1.CustomResourceValidation{OpenAPIV3Schema: openapi.SchemaFrom(reflect.TypeFor[v1alpha1.Flight]())},
+					AdditionalPrinterColumns: []apiextensionsv1.CustomResourceColumnDefinition{
+						{
+							Name:     "flight",
+							Type:     "string",
+							JSONPath: ".spec.wasmUrl",
+						},
+						{
+							Name:     "cluster_access",
+							Type:     "boolean",
+							JSONPath: ".spec.clusterAccess",
+						},
+						{
+							Name:     "fix_drift_interval",
+							Type:     "string",
+							Format:   "date",
+							JSONPath: ".spec.fixDriftInterval",
+						},
+					},
+				},
+			},
+		},
+	}
+
+	flightResource, err := internal.ToUnstructured(flightDef)
+	if err != nil {
+		return fmt.Errorf("failed to convert airway crd to its unstructured representation: %w", err)
+	}
+
+	if err := client.ApplyResources(
+		ctx,
+		[]*unstructured.Unstructured{flightResource, airwayResource},
+		k8s.ApplyResourcesOpts{ApplyOpts: forceful},
+	); err != nil {
 		return fmt.Errorf("failed to apply airway crd: %w", err)
 	}
